requirement.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
jieba==0.42.1
snownlp==0.12.3
requests==2.32.5
openai==2.15.0
python-dotenv==1.0.0
langgraph>=0.2.0
langchain-openai>=0.2.0
